Add explicit URLs to dropdown toggles in navbar.
This allows users to be able to navigate even if JavaScript fails to
load (which has been happening more than usual recently), and by
extension, if JavaScript is disabled or unsupported by the browser.

The previous behavior would have implicitly taken the user to the home
page, whereas this at least gets them closer to their destination.